refactor: move helper functions from bots to other files
Instead of using the nifty `_` for private functions move the helper functions to other files.
Instead of making a big `helper` file I decided to put things in semantically interesting files:
* repo_bots - control the interaction with configuration on a repo level
* owner_bots - control the interaction with configuration on a owner level
* public_bots - control the interaction with configuration on an installation level (YAML configured)

Usually the analysis follows that order of priority.
One small issue was that token mapping would get the admin_bo (redoing the search for the appropriate bot),
and that would cause a circular import, so I just deciede to re-use the search value and pass the admin bot token
as an arg.

Many changes in tests and import paths. In particular some tests became obsolete from the interface change.
Because the token_type_mapping function doesn't do the search anymore, tests that depended on that broke down.
So I removed them, but their "testing" ability is captured in new tests on `services/bots/tests/test_bots.py`
